backend/requirements.txt
python-pptx
pydantic
python-dotenv
phidata 
